feat(overview): drop classic mode, make enriched the only layout
Retires the Classic Overview dial. There is one Overview layout now — the
former "enriched", which is what users actually want — so the toggle was
removed entirely and the Classic dial deleted.

## Migration (written + tested FIRST)

v7 → v8 hop coverage:
- v4 → v8 (four hops)
- v5 → v8 (three hops)
- v6 → v8 (two hops)
- v7 → v8 (one hop)
- v8 pass-through (plus stray-overviewMode + stray-Lanes-view coerce guards)

Plus debug-log presence + absence assertions: stepV7toV8 logs at
console.debug when a payload carried overviewMode (returning Classic
user's breadcrumb), silent for payloads without the field.

Intermediate LegacySettings type carries overviewMode through v5→v6→v7
so stepV7toV8 can both see the value (for the log) and strip it.
readSettingsField now returns the clean v8 Settings; readLegacySettingsField
is the intermediate reader that preserves overviewMode for the chain.

## Deletions

- src/lib/components/ChronographDial.svelte (old classic dial)
- Classic-only derivations in OverviewView: classicVerdict, classicVerdictStyle,
  overCount, totalSamples, liveMedianParts, handleClassicDrill
- Classic-only CSS blocks: diagnosis-strip, diag-*, metrics-triptych,
  metric-*, .overview.classic, .dial-slot
- isEnriched gate on every $effect/$derived (baseline, events, trace,
  samples, nowTimer)
- Classic-only imports: fmt, fmtParts, fmtCount, overviewVerdict,
  VERDICT_STYLES, ChronographDialV2

## Renames

- ChronographDialV2.svelte → ChronographDial.svelte (git mv)
  Component name updated in OverviewView import. File-header comment
  rewritten; the "sibling to Classic" framing is now historical.

## Type narrowing

- Removed OverviewMode type from types.ts
- Dropped overviewMode from Settings interface + DEFAULT_SETTINGS
- PersistedSettings.version now 2..8

## Non-negotiables

- No silent data loss: stepV7toV8 logs the drop; LegacySettings keeps
  the value intact through the intermediate chain so the log has
  something to report.
- Bundle delta NEGATIVE: JS 62.31 → 60.38 KB gzip (-1.93 KB), CSS
  10.35 → 9.84 KB gzip (-0.51 KB). Combined -2.44 KB gzip.
- Hop tests green before any deletion began (commit history enforces).
- Full suite: 571/571 (+4 new hop tests), typecheck clean, lint clean.
