+++ tests/unit/persistence.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect, beforeEach } from 'vitest';
+import { describe, it, expect, beforeEach, vi } from 'vitest';
 import { loadPersistedSettings, saveSettings, migrateSettings } from '../../src/lib/utils/persistence';
 import type { PersistedSettings } from '../../src/lib/types';
 
@@ -20,7 +20,7 @@ describe('persistence', () => {
     expect(loadPersistedSettings()).toBeNull();
   });
 
-  it('round-trips v3 settings into current v7 shape', () => {
+  it('round-trips v3 settings into current v8 shape', () => {
     const settings = {
       version: 3,
       endpoints: [{ url: 'https://example.com', enabled: true }],
@@ -29,13 +29,13 @@ describe('persistence', () => {
     };
     saveSettings(settings as unknown as PersistedSettings);
     const loaded = loadPersistedSettings();
-    expect(loaded?.version).toBe(7);
+    expect(loaded?.version).toBe(8);
     expect(loaded?.endpoints[0]?.url).toBe('https://example.com');
     expect(loaded?.settings.burstRounds).toBe(50);
     expect(loaded?.settings.monitorDelay).toBe(1000);
     expect(loaded?.settings.healthThreshold).toBe(120);
-    // v5→v6 seeds overviewMode; default is 'classic' until Enriched is promoted.
-    expect(loaded?.settings.overviewMode).toBe('classic');
+    // overviewMode was retired at v8; migration strips the field entirely.
+    expect('overviewMode' in (loaded?.settings ?? {})).toBe(false);
   });
 
   it('returns null for corrupt data', () => {
@@ -53,23 +53,23 @@ describe('persistence', () => {
     };
     localStorageMock.setItem('chronoscope_v2_settings', JSON.stringify(settings));
     const loaded = loadPersistedSettings();
-    expect(loaded?.version).toBe(7);
+    expect(loaded?.version).toBe(8);
     expect(localStorageMock.getItem('chronoscope_v2_settings')).toBeNull();
     expect(localStorageMock.getItem('chronoscope_settings')).not.toBeNull();
   });
 
-  it('migrates v1 data all the way to v7', () => {
+  it('migrates v1 data all the way to v8', () => {
     const v1Data = { version: 1, endpoints: [{ url: 'https://example.com' }] };
     const migrated = migrateSettings(v1Data);
-    expect(migrated?.version).toBe(7);
+    expect(migrated?.version).toBe(8);
     expect(migrated?.settings.burstRounds).toBe(50);
     expect(migrated?.settings.monitorDelay).toBe(1000);
     expect(migrated?.settings.healthThreshold).toBe(120);
-    expect(migrated?.settings.overviewMode).toBe('classic');
+    expect('overviewMode' in (migrated?.settings ?? {})).toBe(false);
     expect(migrated?.ui.activeView).toBe('overview');
   });
 
-  it('migrates v2 data to v7 with old delay as monitorDelay', () => {
+  it('migrates v2 data to v8 with old delay as monitorDelay', () => {
     const v2Data = {
       version: 2,
       endpoints: [{ url: 'https://example.com', enabled: true }],
@@ -77,18 +77,17 @@ describe('persistence', () => {
       ui: { expandedCards: [], activeView: 'split' },
     };
     const migrated = migrateSettings(v2Data);
-    expect(migrated?.version).toBe(7);
+    expect(migrated?.version).toBe(8);
     expect(migrated?.settings.monitorDelay).toBe(500);
     expect(migrated?.settings.burstRounds).toBe(50);
     expect(migrated?.settings.delay).toBe(0);
     expect(migrated?.settings.healthThreshold).toBe(120);
-    expect(migrated?.settings.overviewMode).toBe('classic');
-    // v2 'split' → v5 'lanes' → v7 'overview'.
+    // v2 'split' → v5 'lanes' → v7 'overview' → v8 still 'overview'.
     expect(migrated?.ui.activeView).toBe('overview');
   });
 
-  describe('v4 → v5 → v6 → v7 migration (chain)', () => {
-    it('seeds healthThreshold default + overviewMode on a real v4 payload', () => {
+  describe('v4 → v5 → v6 → v7 → v8 migration (chain)', () => {
+    it('seeds healthThreshold default on a real v4 payload', () => {
       const v4: unknown = {
         version: 4,
         endpoints: [{ url: 'https://cloudflare-dns.com', enabled: true }],
@@ -104,10 +103,10 @@ describe('persistence', () => {
         ui: { expandedCards: ['ep-1'], activeView: 'split' },
       };
       const migrated = migrateSettings(v4);
-      expect(migrated?.version).toBe(7);
+      expect(migrated?.version).toBe(8);
       expect(migrated?.settings.healthThreshold).toBe(120);
-      expect(migrated?.settings.overviewMode).toBe('classic');
       expect(migrated?.settings.region).toBe('north-america');
+      expect('overviewMode' in (migrated?.settings ?? {})).toBe(false);
     });
 
     it('rewrites deprecated activeView values through Lanes to "overview" (v7 collapse)', () => {
@@ -160,8 +159,8 @@ describe('persistence', () => {
     });
   });
 
-  describe('v5 → v7 migration', () => {
-    it('seeds overviewMode=classic on a real v5 payload, preserves everything else', () => {
+  describe('v5 → v8 migration', () => {
+    it('preserves modern v5 fields and strips overviewMode at v8', () => {
       const v5: unknown = {
         version: 5,
         endpoints: [{ url: 'https://a.example', enabled: true }],
@@ -179,8 +178,8 @@ describe('persistence', () => {
         },
       };
       const migrated = migrateSettings(v5);
-      expect(migrated?.version).toBe(7);
-      expect(migrated?.settings.overviewMode).toBe('classic');
+      expect(migrated?.version).toBe(8);
+      expect('overviewMode' in (migrated?.settings ?? {})).toBe(false);
       // Existing v5 fields survive intact.
       expect(migrated?.settings.healthThreshold).toBe(180);
       expect(migrated?.settings.region).toBe('europe');
@@ -191,10 +190,9 @@ describe('persistence', () => {
       expect(migrated?.ui.expandedCards).toEqual(['ep-1']);
     });
 
-    it('preserves an existing overviewMode value when present on v5 (forward-compat guard)', () => {
-      // Defensive: if a future/debug tool injected overviewMode into a v5
-      // payload, the migration should carry the value forward rather than
-      // overwrite with the default.
+    it('drops a forward-written overviewMode on v5 at the v8 boundary', () => {
+      // A future/debug tool may have injected overviewMode into a v5 payload;
+      // the chain carries it forward to v6/v7, then stepV7toV8 strips it.
       const v5 = {
         version: 5,
         endpoints: [],
@@ -206,10 +204,10 @@ describe('persistence', () => {
         ui: { expandedCards: [], activeView: 'overview' },
       };
       const migrated = migrateSettings(v5);
-      expect(migrated?.settings.overviewMode).toBe('enriched');
+      expect('overviewMode' in (migrated?.settings ?? {})).toBe(false);
     });
 
-    it('coerces a garbage overviewMode value to the default "classic"', () => {
+    it('ignores a garbage overviewMode on v5 (coerced in the chain, dropped at v8)', () => {
       const v5 = {
         version: 5,
         endpoints: [],
@@ -221,19 +219,65 @@ describe('persistence', () => {
         ui: { expandedCards: [], activeView: 'overview' },
       };
       const migrated = migrateSettings(v5);
-      expect(migrated?.settings.overviewMode).toBe('classic');
+      expect('overviewMode' in (migrated?.settings ?? {})).toBe(false);
     });
   });
 
-  describe('v7 pass-through', () => {
-    it('round-trips a v7 payload unchanged (already current)', () => {
-      const v7: PersistedSettings = {
-        version: 7,
+  describe('v7 → v8 hop', () => {
+    it('drops overviewMode from a real v7 payload and debug-logs the transition', () => {
+      const debugSpy = vi.spyOn(console, 'debug').mockImplementation(() => undefined);
+      try {
+        const v7 = {
+          version: 7,
+          endpoints: [],
+          settings: {
+            timeout: 5000, delay: 0, burstRounds: 50, monitorDelay: 1000,
+            cap: 0, corsMode: 'no-cors', healthThreshold: 120,
+            overviewMode: 'enriched',
+          },
+          ui: { expandedCards: [], activeView: 'overview' },
+        };
+        const migrated = migrateSettings(v7);
+        expect(migrated?.version).toBe(8);
+        expect('overviewMode' in (migrated?.settings ?? {})).toBe(false);
+        expect(debugSpy).toHaveBeenCalledWith(
+          expect.stringMatching(/v8 migration: settings\.overviewMode 'enriched' retired/),
+        );
+      } finally {
+        debugSpy.mockRestore();
+      }
+    });
+
+    it('v7 payload without overviewMode passes through quietly', () => {
+      const debugSpy = vi.spyOn(console, 'debug').mockImplementation(() => undefined);
+      try {
+        const v7 = {
+          version: 7,
+          endpoints: [],
+          settings: {
+            timeout: 5000, delay: 0, burstRounds: 50, monitorDelay: 1000,
+            cap: 0, corsMode: 'no-cors', healthThreshold: 120,
+          },
+          ui: { expandedCards: [], activeView: 'live' },
+        };
+        const migrated = migrateSettings(v7);
+        expect(migrated?.version).toBe(8);
+        expect(migrated?.ui.activeView).toBe('live');
+        expect(debugSpy).not.toHaveBeenCalled();
+      } finally {
+        debugSpy.mockRestore();
+      }
+    });
+  });
+
+  describe('v8 pass-through', () => {
+    it('round-trips a v8 payload unchanged', () => {
+      const v8: PersistedSettings = {
+        version: 8,
         endpoints: [{ url: 'https://a.example', enabled: true }],
         settings: {
           timeout: 5000, delay: 0, burstRounds: 50, monitorDelay: 1000,
           cap: 0, corsMode: 'no-cors', healthThreshold: 200,
-          overviewMode: 'enriched',
         },
         ui: {
           expandedCards: [],
@@ -243,41 +287,26 @@ describe('persistence', () => {
           terminalFilters: ['timeout', 'error'],
         },
       };
-      saveSettings(v7);
+      saveSettings(v8);
       const loaded = loadPersistedSettings();
-      expect(loaded).toEqual(v7);
+      expect(loaded).toEqual(v8);
     });
 
-    it('accepts both valid overviewMode values through the v6→v7 step', () => {
-      for (const mode of ['classic', 'enriched'] as const) {
-        const v6: PersistedSettings = {
-          version: 6,
-          endpoints: [],
-          settings: {
-            timeout: 5000, delay: 0, burstRounds: 50, monitorDelay: 1000,
-            cap: 0, corsMode: 'no-cors', healthThreshold: 120,
-            overviewMode: mode,
-          },
-          ui: { expandedCards: [], activeView: 'overview' },
-        };
-        const migrated = migrateSettings(v6);
-        expect(migrated?.settings.overviewMode).toBe(mode);
-      }
-    });
-
-    it('rejects a corrupted overviewMode by falling back to "classic" (v7 still guards)', () => {
+    it('a stray overviewMode in a v8 payload is silently ignored by the reader', () => {
+      // Not a real migration path — a hand-edited or corrupted v8 payload
+      // shouldn't let the retired field leak back into the app.
       const bad = {
-        version: 7,
+        version: 8,
         endpoints: [],
         settings: {
           timeout: 5000, delay: 0, burstRounds: 50, monitorDelay: 1000,
           cap: 0, corsMode: 'no-cors', healthThreshold: 120,
-          overviewMode: 42,   // wrong type — shouldn't happen, but we guard
+          overviewMode: 'classic',
         },
         ui: { expandedCards: [], activeView: 'overview' },
       };
       const migrated = migrateSettings(bad);
-      expect(migrated?.settings.overviewMode).toBe('classic');
+      expect('overviewMode' in (migrated?.settings ?? {})).toBe(false);
     });
   });
 
@@ -288,7 +317,7 @@ describe('persistence', () => {
         endpoints: [{ url: 'https://example.com', enabled: true }],
         settings: {
           timeout: 5000, delay: 0, burstRounds: 50, monitorDelay: 1000,
-          cap: 0, corsMode: 'no-cors', healthThreshold: 120, overviewMode: 'classic',
+          cap: 0, corsMode: 'no-cors', healthThreshold: 120,
         },
         ui: { expandedCards: [], activeView: 'overview' },
         unknownFutureField: 'ignored',
